requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=3.13.0
django-filter>=22.1
django-solo>=2.0.0
Pillow>=9.0.0
python-decouple>=3.6
gunicorn>=20.1.0
whitenoise>=6.0.0
psycopg2-binary>=2.9.0
dj-database-url>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
# JWT auth
djangorestframework-simplejwt>=5.3.1